requirements.txt
flask
ultralytics
pillow
requests
torch